Move DevOptionsSettingsFragment to a new file
